refactor: Auth, User API 및 기능 분리 #99
