+++ BE/ytt/src/main/java/com/example/ytt/domain/user/auth/security/CustomLogoutFilter.java
@@ -31,7 +31,7 @@ public class CustomLogoutFilter extends GenericFilterBean {
 
         //path and method verify
         String requestUri = request.getRequestURI();
-        if (!requestUri.matches("^\\/user\\/logout$")) {
+        if (!requestUri.matches("^\\/auth\\/logout$")) {
             filterChain.doFilter(request, response);
             return;
         }
